docs: Domain 패키지 생성및 클래스 위치 수정
