+++ src/main/java/com/sprint/mission/discodeit/domain/dto/AlarmDTO.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.domain.dto;
 
 import java.util.UUID;
