+++ src/main/java/com/sprint/mission/discodeit/domain/dto/ChannelCreateDTO.java
@@ -1,6 +1,6 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.domain.dto;
 
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.domain.entity.ChannelType;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
