+++ src/main/java/com/sprint/mission/discodeit/domain/dto/UserCreateDTO.java
@@ -1,0 +1,17 @@
+package com.sprint.mission.discodeit.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UserCreateDTO {
+	private String username;
+	private String email;
+	private String password;
+
+}
