+++ src/main/java/com/sprint/mission/discodeit/domain/entity/Channel.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.domain.entity;
 
 import java.io.Serial;
 import java.io.Serializable;
