+++ src/main/java/com/sprint/mission/discodeit/domain/entity/ChannelUser.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.domain.entity;
 
 import java.util.UUID;
