--- src/main/java/com/sprint/mission/discodeit/dto/MessageCreateDTO.java
@@ -1,18 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.util.UUID;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MessageCreateDTO {
-	private String content;
-	private UUID channelId;
-	private java.util.UUID userId;
-}
