+++ src/main/java/com/sprint/mission/discodeit/repository/BinaryContentRepository.java
@@ -3,7 +3,7 @@ package com.sprint.mission.discodeit.repository;
 import java.util.List;
 import java.util.UUID;
 
-import com.sprint.mission.enums.ContentType;
+import com.sprint.mission.discodeit.domain.enums.ContentType;
 
 public interface BinaryContentRepository {
 	public BinaryContentRepository create(byte[] content, Long size, ContentType contentType, String filename);
