+++ src/main/java/com/sprint/mission/discodeit/repository/MessageRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.domain.entity.Message;
 
 public interface MessageRepository {
 	Message save(Message message);
