+++ src/main/java/com/sprint/mission/discodeit/repository/UserRepository.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Repository;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.domain.entity.User;
 
 @Repository
 public interface UserRepository {
