+++ src/main/java/com/sprint/mission/discodeit/repository/UserStatusRepository.java
@@ -3,7 +3,7 @@ package com.sprint.mission.discodeit.repository;
 import java.util.List;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.entity.UserStatus;
+import com.sprint.mission.discodeit.domain.entity.UserStatus;
 
 public interface UserStatusRepository {
