+++ src/main/java/com/sprint/mission/discodeit/repository/file/FileChannelRepository.java
@@ -15,8 +15,8 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Repository;
 
-import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.domain.entity.Channel;
+import com.sprint.mission.discodeit.domain.entity.Message;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 
 @Repository
