+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFChannelRepository.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.domain.entity.Channel;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 
 public class JCFChannelRepository implements ChannelRepository {
