+++ src/main/java/com/sprint/mission/discodeit/repository/jcf/JCFUserRepository.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.domain.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 
 public class JCFUserRepository implements UserRepository {
