+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -3,9 +3,9 @@ package com.sprint.mission.discodeit.service;
 import java.util.List;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.dto.ChannelCreateDTO;
-import com.sprint.mission.discodeit.dto.ChannelUpdateDTO;
-import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.domain.dto.ChannelCreateDTO;
+import com.sprint.mission.discodeit.domain.dto.ChannelUpdateDTO;
+import com.sprint.mission.discodeit.domain.entity.Channel;
 
 public interface ChannelService {
 	public Channel create(ChannelCreateDTO dto);
