+++ src/main/java/com/sprint/mission/discodeit/service/jsf/JCFChannelService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
-import com.sprint.mission.discodeit.dto.ChannelCreateDTO;
-import com.sprint.mission.discodeit.dto.ChannelUpdateDTO;
-import com.sprint.mission.discodeit.entity.Channel;
-import com.sprint.mission.discodeit.entity.ChannelType;
+import com.sprint.mission.discodeit.domain.dto.ChannelCreateDTO;
+import com.sprint.mission.discodeit.domain.dto.ChannelUpdateDTO;
+import com.sprint.mission.discodeit.domain.entity.Channel;
+import com.sprint.mission.discodeit.domain.entity.ChannelType;
 import com.sprint.mission.discodeit.repository.jcf.JCFChannelRepository;
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
